My Jetpack: Lift TS types up for reusability.
