+++ projects/packages/my-jetpack/_inc/components/module-status/index.tsx
@@ -1,6 +1,6 @@
 import { Badge } from '@automattic/ui';
 import { __ } from '@wordpress/i18n';
-import { MyJetpackModule } from '../types';
+import { MyJetpackModule } from '../../types';
 
 import '@automattic/ui/style.css';
