--- projects/packages/my-jetpack/_inc/components/types.ts
@@ -1,9 +1,0 @@
-export type MyJetpackModule = {
-	available: boolean;
-	module: string;
-	name: string;
-	activated: boolean;
-	description: string;
-	long_description: string;
-	search_terms: string;
-};
